Split out getTermIdsFromEdits
