+++ client/state/posts/utils/get-term-ids-from-edits.js
@@ -1,0 +1,48 @@
+/**
+ * External dependencies
+ */
+import { isEmpty, isPlainObject, map, mapValues, reduce, toArray } from 'lodash';
+
+/**
+ * Takes existing term post edits and updates the `terms_by_id` attribute
+ *
+ * @param  {object}    post  object of post edits
+ * @returns {object}          normalized post edits
+ */
+export function getTermIdsFromEdits( post ) {
+	if ( ! post || ! post.terms ) {
+		return post;
+	}
+
+	// Filter taxonomies that are set as arrays ( i.e. tags )
+	// This can be detected by an array of strings vs an array of objects
+	const taxonomies = reduce(
+		post.terms,
+		( prev, taxonomyTerms, taxonomyName ) => {
+			// Ensures we are working with an array
+			const termsArray = toArray( taxonomyTerms );
+			if ( termsArray && termsArray.length && ! isPlainObject( termsArray[ 0 ] ) ) {
+				return prev;
+			}
+
+			prev[ taxonomyName ] = termsArray;
+			return prev;
+		},
+		{}
+	);
+
+	if ( isEmpty( taxonomies ) ) {
+		return post;
+	}
+
+	return {
+		...post,
+		terms_by_id: mapValues( taxonomies, taxonomy => {
+			const termIds = map( taxonomy, 'ID' );
+
+			// Hack: qs omits empty arrays in wpcom.js request, which prevents
+			// removing all terms for a given taxonomy since the empty array is not sent to the API
+			return termIds.length ? termIds : null;
+		} ),
+	};
+}
